+++ src/Exceptionless.Core/Models/UsageInfo.cs
@@ -9,7 +9,7 @@ public record UsageInfo
     public int Blocked { get; set; }
     public int Discarded { get; set; }
     public int TooBig { get; set; }
-    public long Deleted { get; set; }
+    public int Deleted { get; set; }
 }
 
 public record UsageHourInfo
@@ -19,7 +19,7 @@ public record UsageHourInfo
     public int Blocked { get; set; }
     public int Discarded { get; set; }
     public int TooBig { get; set; }
-    public long Deleted { get; set; }
+    public int Deleted { get; set; }
 }
 
 public record UsageInfoResponse
